requirements.txt
opencv-python~=4.7.0.72
numpy~=1.24.2
Pillow~=9.4.0
pyzbar~=0.1.9
fastapi~=0.94.1
starlette~=0.26.1
pydantic~=1.10.6
uvicorn~=0.21.1
loguru
python-multipart
python-dotenv
jinja2